Configure custom authentication provider
